Move stuff to DeviceService
- Its easier if everything is together :)
