--- server-services-interfaces/src/main/java/ar/edu/itba/iot/carne_iot/server/services/DevicePairingService.java
@@ -1,24 +1,0 @@
-package ar.edu.itba.iot.carne_iot.server.services;
-
-/**
- * Defines behaviour of the service
- * in charge of managing {@link ar.edu.itba.iot.carne_iot.server.models.Device} pairing.
- */
-public interface DevicePairingService {
-
-    /**
-     * Constant indicating a role for a {@link ar.edu.itba.iot.carne_iot.server.models.Device}
-     * (i.e the one operating with the server is a {@link ar.edu.itba.iot.carne_iot.server.models.Device}).
-     */
-    String DEVICE_ROLE = "DEVICE";
-
-    /**
-     * Performs pairing of {@link ar.edu.itba.iot.carne_iot.server.models.Device}s.
-     *
-     * @param ownerId  The id of the {@link ar.edu.itba.iot.carne_iot.server.models.User}
-     *                 owning the {@link ar.edu.itba.iot.carne_iot.server.models.Device}.
-     * @param deviceId The id of the {@link ar.edu.itba.iot.carne_iot.server.models.Device} to be paired.
-     * @return The token generated by the pairing process.
-     */
-    String pair(long ownerId, long deviceId);
-}
