+++ server-services-interfaces/src/main/java/ar/edu/itba/iot/carne_iot/server/services/DeviceService.java
@@ -15,6 +15,12 @@ import java.util.Optional;
 public interface DeviceService {
 
     /**
+     * Constant indicating a role for a {@link ar.edu.itba.iot.carne_iot.server.models.Device}
+     * (i.e the one operating with the server is a {@link ar.edu.itba.iot.carne_iot.server.models.Device}).
+     */
+    String DEVICE_ROLE = "DEVICE";
+
+    /**
      * List all {@link Device}s stored in the system, in a paginated view.
      *
      * @param pageable The {@link Pageable} object containing pagination stuff.
@@ -93,6 +99,16 @@ public interface DeviceService {
      * @param deviceId The id of the {@link Device} being unregistered.
      */
     void unregisterDevice(long deviceId);
+
+    /**
+     * Performs pairing of {@link ar.edu.itba.iot.carne_iot.server.models.Device}s.
+     *
+     * @param ownerId  The id of the {@link ar.edu.itba.iot.carne_iot.server.models.User}
+     *                 owning the {@link ar.edu.itba.iot.carne_iot.server.models.Device}.
+     * @param deviceId The id of the {@link ar.edu.itba.iot.carne_iot.server.models.Device} to be paired.
+     * @return The token generated by the pairing process.
+     */
+    String pair(long ownerId, long deviceId);
 
     /**
      * Makes a {@link Device} start cooking (i.e change its state to {@link Device.State#ACTIVE}).
