+++ server-services/src/main/java/ar/edu/itba/iot/carne_iot/server/services/DeviceServiceImpl.java
@@ -7,12 +7,15 @@ import ar.edu.itba.iot.carne_iot.server.exceptions.CustomIllegalStateException;
 import ar.edu.itba.iot.carne_iot.server.exceptions.NoSuchEntityException;
 import ar.edu.itba.iot.carne_iot.server.models.Device;
 import ar.edu.itba.iot.carne_iot.server.models.DeviceRegistration;
+import ar.edu.itba.iot.carne_iot.server.models.Session;
 import ar.edu.itba.iot.carne_iot.server.models.User;
 import ar.edu.itba.iot.carne_iot.server.persistence.daos.DeviceDao;
 import ar.edu.itba.iot.carne_iot.server.persistence.daos.DeviceRegistrationDao;
+import ar.edu.itba.iot.carne_iot.server.persistence.daos.SessionDao;
 import ar.edu.itba.iot.carne_iot.server.persistence.daos.UserDao;
 import ar.edu.itba.iot.carne_iot.server.persistence.query_helpers.DeviceQueryHelper;
 import ar.edu.itba.iot.carne_iot.server.persistence.query_helpers.DeviceRegistrationQueryHelper;
+import ar.edu.itba.iot.carne_iot.server.security.JwtTokenGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,6 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.Collections;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -34,6 +38,11 @@ import java.util.function.Function;
 @Service
 @Transactional(readOnly = true)
 public class DeviceServiceImpl implements DeviceService, UniqueViolationExceptionThrower {
+
+    /**
+     * Amount of tries to perform the pairing process.
+     */
+    private static final int MAX_TRIES = 10;
 
     /**
      * DAO for managing {@link User}s data.
@@ -60,16 +69,29 @@ public class DeviceServiceImpl implements DeviceService, UniqueViolationExceptio
      */
     private final DeviceRegistrationQueryHelper deviceRegistrationQueryHelper;
 
+    /**
+     * A {@link JwtTokenGenerator} to create tokens when login is performed.
+     */
+    private final JwtTokenGenerator jwtTokenGenerator;
+
+    /**
+     * DAO for retrieving {@link Session}s data.
+     */
+    private final SessionDao sessionDao;
+
 
     @Autowired
     public DeviceServiceImpl(UserDao userDao, DeviceDao deviceDao, DeviceRegistrationDao deviceRegistrationDao,
                              DeviceQueryHelper deviceQueryHelper,
-                             DeviceRegistrationQueryHelper deviceRegistrationQueryHelper) {
+                             DeviceRegistrationQueryHelper deviceRegistrationQueryHelper,
+                             JwtTokenGenerator jwtTokenGenerator, SessionDao sessionDao) {
         this.userDao = userDao;
         this.deviceDao = deviceDao;
         this.deviceRegistrationDao = deviceRegistrationDao;
         this.deviceQueryHelper = deviceQueryHelper;
         this.deviceRegistrationQueryHelper = deviceRegistrationQueryHelper;
+        this.jwtTokenGenerator = jwtTokenGenerator;
+        this.sessionDao = sessionDao;
     }
 
 
@@ -176,6 +198,39 @@ public class DeviceServiceImpl implements DeviceService, UniqueViolationExceptio
                     reg.inactivate();
                     deviceRegistrationDao.save(reg);
                 });
+    }
+
+    @Override
+    @Transactional
+    @PreAuthorize("@devicePermissionProvider.isOwnerOrAdmin(#deviceId)")
+    public String pair(long ownerId, long deviceId) {
+        final User user = userDao.findById(ownerId).orElseThrow(NoSuchEntityException::new);
+        final Device device = deviceDao.findById(deviceId).orElseThrow(NoSuchEntityException::new);
+
+        // Verify the device is registered
+        if (!deviceRegistrationDao.existsByDeviceAndActiveTrue(device)) {
+            throw new CustomIllegalStateException(OPERATION_OVER_UNREGISTERED_DEVICE);
+        }
+
+        // Try to create the token...
+        boolean validToken = false;
+        int tries = 0;
+        JwtTokenGenerator.TokenAndSessionContainer container = null;
+        while (!validToken && tries < MAX_TRIES) {
+            container = jwtTokenGenerator.generateDeviceToken(user, device);
+            validToken = !sessionDao.existsByOwnerAndJti(user, container.getJti());
+            tries++;
+        }
+        if (tries >= MAX_TRIES) {
+            throw new RuntimeException("Could not create a session after " + MAX_TRIES + "tries");
+        }
+
+        // Store token in order to pass the
+        Objects.requireNonNull(container, "The container was not initialized correctly");
+        final Session session = new Session(user, container.getJti());  // Actually not a session, but whatever
+        sessionDao.save(session);
+
+        return container.getToken();
     }
 
     @Override
